Commit Aula 3, autenticação

+++ server/src/routes/memories.ts
@@ -2,23 +2,36 @@ import { FastifyInstance } from "fastify";
 import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 export async function memoriesRoutes(app: FastifyInstance){
-
-    app.get('/memories', async () => {
-        const memories = await prisma.memory.findMany({
-        orderBy:{
-            createdAt:"asc"
-        }
-        })
-       return memories.map((memory) => {
-        return{
-        id: memory.id,
-        coverUrl: memory.coverUrl,
-        excerpt: memory.content.substring(0,115).concat('...'),
-       }
-       })
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
     })
 
-    app.get('/memories/:id', async (request) => {      
+    app.get('/memories', async (request) => {
+        await request.jwtVerify()
+
+        const {} = request.user
+
+
+        app.get('/memories', async () => {
+            const memories = await prisma.memory.findMany({
+                where: {
+                    userId: request.user.sub,
+                },
+            orderBy:{
+                createdAt:"asc"
+            },
+            })
+        return memories.map((memory) => {
+            return{
+            id: memory.id,
+            coverUrl: memory.coverUrl,
+            excerpt: memory.content.substring(0,115).concat('...'),
+        }
+        })
+        })
+    })
+    
+    app.get('/memories/:id', async (request, reply) => {      
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -29,7 +42,9 @@ export async function memoriesRoutes(app: FastifyInstance){
                 id,
             },
         })
-
+        if (!memory.isPublic && memory.userId != request.user.sub){
+            return reply.status(401).send()
+        }
         return memory
     })
 
@@ -46,15 +61,13 @@ export async function memoriesRoutes(app: FastifyInstance){
                 content,
                 coverUrl,
                 isPublic,
-                userId:''
+                userId: request.user.sub
             }
         })
         return memory
     })
 
-
-
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -68,7 +81,17 @@ export async function memoriesRoutes(app: FastifyInstance){
 
         const { content ,coverUrl, isPublic } = bodySchema.parse(request.body)
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
+
+         memory = await prisma.memory.update({
             where:{
                 id,
             },
@@ -81,14 +104,23 @@ export async function memoriesRoutes(app: FastifyInstance){
         return memory
     })
 
-
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
          const paramsSchema = z.object({
             id: z.string().uuid(),
         })
         const { id } = paramsSchema.parse(request.params)
 
-        const memory = await prisma.memory.delete({
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
+
+        await prisma.memory.delete({
             where: {
                 id,
             },
